+++ idp-commons/src/main/java/de/gematik/idp/IdpConstants.java
@@ -39,15 +39,9 @@ public final class IdpConstants {
   public static final String FED_AUTH_ENDPOINT = "/auth";
   public static final String FED_AUTH_APP_ENDPOINT = "/auth/app";
 
-  public static final String FED_LIST_ENDPOINT = "/federation_list";
-
   public static final String SEKTORAL_IDP_AUTHORIZATION_ENDPOINT = "/authorization";
   public static final String ENTITY_STATEMENT_ENDPOINT = "/.well-known/openid-federation";
-
   public static final String IDP_LIST_ENDPOINT = "/.well-known/idp_list";
-
-  public static final String FEDMASTER_FEDERATION_FETCH_ENDPOINT = "/federation_fetch_endpoint";
-
   public static final String ENTITY_STATEMENT_TYP = "entity-statement+jwt";
 
   public static final String OPENID = "openid";