+++ idp-commons/src/main/java/de/gematik/idp/token/IdTokenBuilder.java
@@ -17,7 +17,6 @@
 package de.gematik.idp.token;
 
 import static de.gematik.idp.IdpConstants.EREZEPT;
-import static de.gematik.idp.IdpConstants.OID_VERSICHERTER;
 import static de.gematik.idp.field.ClaimName.ACCESS_TOKEN_HASH;
 import static de.gematik.idp.field.ClaimName.AUDIENCE;
 import static de.gematik.idp.field.ClaimName.AUTHENTICATION_CLASS_REFERENCE;
@@ -38,6 +37,7 @@ import static de.gematik.idp.field.ClaimName.ORGANIZATION_NAME;
 import static de.gematik.idp.field.ClaimName.PROFESSION_OID;
 import static de.gematik.idp.field.ClaimName.SUBJECT;
 import static de.gematik.idp.field.ClaimName.TYPE;
+import static de.gematik.idp.token.TokenBuilderUtil.addDisplayNameToBodyClaims;
 import static de.gematik.idp.token.TokenBuilderUtil.buildSubjectClaim;
 
 import de.gematik.idp.IdpConstants;
@@ -96,18 +96,10 @@ public class IdTokenBuilder {
     final Optional<Object> nonceInAuthCode = authenticationToken.getBodyClaim(NONCE);
     nonceInAuthCode.ifPresent(o -> claimsMap.put(NONCE.getJoseName(), o));
     // for erezept in federation
-    final Optional<Object> displayName = authenticationToken.getBodyClaim(DISPLAY_NAME);
-    final Optional<Object> professionOid = authenticationToken.getBodyClaim(PROFESSION_OID);
-    if (displayName.isPresent()) {
-      claimsMap.put(DISPLAY_NAME.getJoseName(), displayName.get());
-    } else if (professionOid.isPresent()
-        && authenticationToken.getScopesBodyClaim().contains(EREZEPT)
-        && professionOid.get().equals(OID_VERSICHERTER)) {
-      claimsMap.put(
-          DISPLAY_NAME.getJoseName(),
-          claimsMap.get(GIVEN_NAME.getJoseName()) + " " + claimsMap.get(FAMILY_NAME.getJoseName()));
+    if (authenticationToken.getScopesBodyClaim().contains(EREZEPT)) {
+      final Optional<Object> displayName = authenticationToken.getBodyClaim(DISPLAY_NAME);
+      addDisplayNameToBodyClaims(displayName, claimsMap);
     }
-
     claimsMap.put(
         AUTHORIZED_PARTY.getJoseName(),
         authenticationToken